feat(deleteAllFiles): return if dir doesnt exist

+++ src/lib/deleteAllFiles.ts
@@ -5,7 +5,11 @@ import path from 'path'
 
 export async function deleteAllFiles(dir: string) {
     // Apaga todos os arquivos que exitirem
-    const files = fs.readdirSync(path.join(process.cwd(), dir));
+    const dirpath = path.join(process.cwd(), dir)
+    
+    if (!fs.existsSync(dirpath)) return;
+
+    const files = fs.readdirSync(dirpath);
     if (files.length > 0){
         for (const file of files) {
             fs.unlinkSync(path.join(dir, file));
